% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{AnomalyPheno-package}
\alias{AnomalyPheno}
\alias{AnomalyPheno-package}
\title{AnomalyPheno: Unsupervised Anomaly Phenotyping for Time Series}
\description{
Provides tools for detecting and classifying anomalies in univariate time
series into interpretable types (phenotypes) using topological data analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/TonyIsFool/anomalypheno}
  \item Report bugs at \url{https://github.com/TonyIsFool/anomalypheno/issues}
}

}
\author{
\strong{Maintainer}: Tony Lu \email{xulunt123@gmail.com}

}
\keyword{internal}
